+++ src/components/z-table/z-table-expanded-row/index.tsx
@@ -13,9 +13,7 @@ export class ZTableExpandedRow {
     return (
       <td colSpan={this.colSpan}>
         <div class="content-container">
-          <div class="inner-content">
-            <slot />
-          </div>
+          <slot />
         </div>
       </td>
     );